fix: add unit tests
